Reinstate non_pod_struct comments.
They're useful.

And put back doc_attrs.
